Add a render as the background for comparison
A render of the original character sheet PDF was made at 300dpi. This
was when the text became relatively clear instead of fuzzy when
displayed as the background of the pdf-pages in the html. By setting
the opacity of the actual background image to 0, the render shows
through and can be compared against added elements. With the opacity
of the actual background image turned back up to 1, the final product
is visible.
opacity of