Fix the justification of the copyright notice
The justification issue turned out to be related to the last line of
text not being justified. I found a trick on
https://kristinlbradley.wordpress.com/2011/09/15/cross-browser-css-justify-last-line-paragraph-text/
to add a 100% wide inline-block after the block that needs
justification to add a pseudo line. This trick allowed justification
in all browsers.

Once this was done I determined the optimal font-size for the
copyright notice is 24.75, but in Safari, wehn zoomed out (100%) this
causes the text to flow onto a second line, so 24.25 is used instead.