Add the name header
Each of the input fields is slightly different in size and slightly
different in margin, which makes automatic flow based sizing and
positioning problematic, so we use absolute.

Even if the padding weren't the case, until CSS has a way to adjust
input elements to the remaining space in a block we'd still need
absolute width.