Switch to create-react-app and add personal data
In order to get fonts to work well inside of the SVGs without using an
embedded data url, we need to have them be inline. Fortunately,
create-react-app has been updated to support Typescript and inline SVGs
without needing to use the Typescript fork, so we could switch to it.

This need was discovered while working on adding the personal data
block, so it was included in this change.