Add the second page to the character sheet
There are some new issues other than the minute placement and sizing
problems seen on the first pages. The second page makes use of the
letter-spacing attribute, which Firefox does not support in SVG (see
https://bugzilla.mozilla.org/show_bug.cgi?id=371787). Some of the
details on the second page have been reproduced correctly, but seem to
be incorrect in the original material. This includes some of the tab
sizes and some of the box sizes. It is an open question whether or not
to "fix" those issues.