Use a constraint system for priority adjustment.
By using constraints, we can assign scores to the various possible
adjustments one constraint at a time. This makes interacting
requirements easier to handle. It also makes selecting the most
appropriate configuration for invalid characters easier.

Note: There is a bug when Pixies (and presumably other naturally
magical metasapients) select Technomancer where the Metatype is set to
A priority while it seems that Magic or Resonance ought to be priority
A. This needs to be investigated, but potentially does not need to be
resolved as they cannot actually be Technomancers.